Reworked JJAX Smart Contract
Now users when minting can add a URI that will be associated with a JSON file on the IPFS protocol. This JSON file will link to two other files on the IFPS protocol being the album cover and the MP3 associated. 

Having trouble deploying to a testnet and therefore being able to test this however it looks good for now.